Finished refactoring quit and fixing goroutine leak
Quit message is now handled like everything else. The quit
messagehandler which contains cleanup of any last open file and shutting
down the gui is scheduled for when gui dispatch thread is ready. In the
meantime the message handling goroutine exits plugging the leak.

The leak was harmless, but fixing it amused me and allowed me to get rid
of annoying messages about special handling of quit.