Improvements to page turning
Reverted to an earlier page turning strategy, but with a new optimization.
Turning a page, a single time, should feel effortless and smooth. Most of all it
should not distract from reading. Before sometimes there was grinding (the
feeling that the computer was struggling). On another attempt it felt jerky or
uneven. With the new setup, the page can always be displayed and then any other
work follows.

Added refreshSpreads command
Reduced time wasted in printLoaded()
Ensure never trying to render a partially loaded spread
Added trace routines to better time function execution