Added cursor hiding, Fixed fullscreen sync, fixed ui resizing, Updated gotk3
Now during periods of inactivity, while in fullscreen, the mouse cursor
will be hidden along with the HUD. The idea is to reduce distractions in
fullscreen.

Fixed fullscreen state sometimes getting out-of-sync. This could happen
if your Window Manager has a mechanism (like a hot-key) for making a
window fullscreen. Internally the toggleFullscreen message was replaced
with setFullscreen. A new window-state-event was added and cbxv commands
now just set the window to Fullscreen or non-Fullscreen. The change in ui
state, regardless of the cause, is caught in the window-state-event and
the setFullscreen message is sent internally to update the model's
fullscreen state.

Fixed ui (header and nav) refusing to resize when no book was open. This
was mainly cosmetic as there's almost nothing for these components to do
when no book is open, but it was ugly.

Updated gotk3 dependency to get the latest improvements.