Add Direction to Layout persistence
Debated about this a long time, but I feel conceptually reading
direction is sort of the most primary layout choice so it should be part
of the Layout that is persisted. Bumped the version number of the Layout
format, but essentially this is just an addition, so the old behavior is
preserved, unless the new data is present in the Layout file, in which
case the program takes advantage of it.